Remove unwanted dependencies, move auth
