+++ src/components/auth/index.tsx
@@ -1,0 +1,3 @@
+import AuthDialog from "./AuthDialog";
+
+export default AuthDialog;
